--- apps/beeai-ui/src/api/mcp-client/index.ts
@@ -1,12 +1,0 @@
-export const MCP_SERVER_URL = new URL('/mcp/sse', location.href);
-export const MCP_EXAMPLE_AGENT_CONFIG = {
-  name: 'example-client',
-  version: '1.0.0',
-};
-export const MCP_EXAMPLE_AGENT_PARAMS = {
-  capabilities: {
-    prompts: {},
-    resources: {},
-    tools: {},
-  },
-};
