--- apps/beeai-ui/src/components/fallbacks/MCPFallback.module.scss
@@ -1,6 +1,0 @@
-.root {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  inline-size: 100%;
-}
