+++ apps/beeai-ui/src/contexts/MCPClient/MCPClientProvider.tsx
@@ -1,9 +1,16 @@
-import { PropsWithChildren } from 'react';
+import { PropsWithChildren, useEffect, useState } from 'react';
 import { MCPClientContext } from './mcp-client-context';
-import { useMCPClient } from '@/api/mcp-client/useMCPClient';
+import { useCreateMCPClient } from '@/api/mcp-client/useCreateMCPClient';
+import { Client as MCPClient } from '@i-am-bee/acp-sdk/client/index.js';
 
 export function MCPClientProvider({ children }: PropsWithChildren) {
-  const client = useMCPClient();
+  const [client, setClient] = useState<MCPClient | null>(null);
+
+  const createClient = useCreateMCPClient();
+
+  useEffect(() => {
+    createClient().then((client) => setClient(client));
+  }, [createClient]);
 
   return <MCPClientContext.Provider value={client}>{children}</MCPClientContext.Provider>;
 }
