+++ apps/beeai-ui/src/contexts/MCPClient/index.ts
@@ -1,6 +1,6 @@
 import { use } from 'react';
 import { MCPClientContext } from './mcp-client-context';
 
-export function useMCPClientContext() {
+export function useMCPClient() {
   return use(MCPClientContext);
 }
