+++ apps/beeai-ui/src/modules/agents/api/queries/useListAgents.ts
@@ -1,4 +1,4 @@
-import { useMCPClientContext } from '@/contexts/MCPClient';
+import { useMCPClient } from '@/contexts/MCPClient';
 import { useQuery } from '@tanstack/react-query';
 import { agentKeys } from '../keys';
 import { Agent, ListAgentsParams } from '../types';
@@ -8,7 +8,7 @@ interface Props {
 }
 
 export function useListAgents({ params }: Props = {}) {
-  const client = useMCPClientContext();
+  const client = useMCPClient();
 
   return useQuery({
     queryKey: agentKeys.list(params),
